requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
awscrt==0.21.1
awsiotsdk==1.22.0
bidict==0.23.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.3
Flask==3.0.3
Flask-SocketIO==5.4.1
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nanopb==0.4.9
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.6
wsproto==1.2.0
